Correction to sort by Date first than grouping using max index later
We were grouping using max Date when creating a df with latest values for each Ticker and Grupo de Compra, which caused issues when 2 operations happened in the same day.

Since we were already sorting by Date in the beginning, we just changed the grouping to consider the max index instead.